Convert global mixins to composables (#281)
* Convert global mixins to composables

Replace the two global Options API mixins with Vue 3 composables under a
new src/composables/ directory:

- useDownload() -> { downloadStager, downloadText } (stateless DOM/blob
  helpers, logic relocated from the download-stager mixin)
- useCopyStager() -> { copyStager }, injecting "snack" internally and
  delegating to copyToClipboard() exactly as before

Migrate all 8 consumers off the mixins via a minimal setup() bridge that
re-exposes the same method names the templates already use, leaving the
Options API components otherwise unchanged. Delete the now-empty
src/mixins/ directory.

Add an .eslintrc.js override disabling import/prefer-default-export for
src/composables/** so composables can follow the idiomatic named-export
(useFoo) convention.

Tighten JSDoc on the composables ({string} params, document the
empty/trailing-slash no-op in downloadStager) and drop a redundant
always-true guard in downloadStager.


* Add CHANGELOG entry for mixin-to-composable conversion


---------
