+++ src/components/agents/AgentTasksList.vue
@@ -68,7 +68,7 @@ import { mapState } from "pinia";
 
 import AgentTasksTable from "@/components/agents/AgentTasksTable.vue";
 import ListPageTop from "@/components/ListPageTop.vue";
-import DownloadMixin from "@/mixins/download-stager";
+import { useDownload } from "@/composables/useDownload";
 import ExpansionPanelFilter from "@/components/tables/ExpansionPanelFilter.vue";
 import ExpansionPanelSearch from "@/components/tables/ExpansionPanelSearch.vue";
 import AdvancedTable from "@/components/tables/AdvancedTable.vue";
@@ -85,7 +85,6 @@ export default {
     AgentTasksTable,
     ListPageTop,
   },
-  mixins: [DownloadMixin],
   inject: ["snack"],
   props: {
     agent: {
@@ -103,6 +102,10 @@ export default {
       type: Boolean,
       default: false,
     },
+  },
+  setup() {
+    const { downloadStager, downloadText } = useDownload();
+    return { downloadStager, downloadText };
   },
   data() {
     return {
