+++ src/components/agents/AgentTasksTable.vue
@@ -367,7 +367,7 @@ import DateTimeDisplay from "@/components/DateTimeDisplay.vue";
 import TooltipButton from "@/components/TooltipButton.vue";
 import TagViewer from "@/components/TagViewer.vue";
 import HeaderMenu from "@/components/HeaderMenu.vue";
-import DownloadMixin from "@/mixins/download-stager";
+import { useDownload } from "@/composables/useDownload";
 import { useApplicationStore } from "@/stores/application-module";
 import * as downloadApi from "@/api/download-api";
 import * as agentTaskApi from "@/api/agent-task-api";
@@ -382,7 +382,6 @@ export default {
     TooltipButton,
     HeaderMenu,
   },
-  mixins: [DownloadMixin],
   inject: ["snack"],
   props: {
     agent: {
@@ -418,6 +417,10 @@ export default {
       type: Boolean,
       default: false,
     },
+  },
+  setup() {
+    const { downloadStager, downloadText } = useDownload();
+    return { downloadStager, downloadText };
   },
   data() {
     return {
